Introduction of a way to sort while outputing to a file (only)
This patch fix #95

You may ask yourself:
  * Why only file?
  * How does it works?

Why only file?

It simple, what you see from the terminal is a live output as long as
we are testing. So, there is no possibility to filter that.

How does it works?

From your side (as a user), you simply have to use the
`--standard-sorting` or the `--hierarchical-sorting` arguments along
with the `--output` argument. The outputed file will be sorted.

From our side, we simply use the sorting logic which was implemented
into PyFunceble. You can read more about it there:
* https://pyfunceble.readthedocs.io/en/latest/code/sort.html
* https://github.com/funilrys/PyFunceble/issues/10