By request
As requested by @funilrys in https://github.com/Ultimate-Hosts-Blacklist/whitelist/pull/105/files#r373779915 I've tried to meet what he wanted. However I would love to see a pre-build of this submission as line breaks needs to be right. I'm also a little in doubt of L286 Is it actually possible to add a line break there?