Introduction of 3 new arguments.
This patch fixes #96.

Indeed, before this patch, it was not possible to give a single file and
leave let our tool prefix it's content with a chosen rule.

From now, 3 new arguments are included.

* `--all`, will read the given file(s) and prefix each rules with the
  `ALL` marker.
* `--reg`, will read the given file(s) and prefix each rules with the
  `REG` marker.
* `--rzd`, will read the given fils(s) and prefix each rules with the
  `RZD` marker.

This let end-user have a better overview of their rules. It also let
end-user have it's own update scheme for the inputs.

WARNING: The `-w` argument does not change.